added ability to determine feedback display distance; refactored other UI-related code
