+++ Assets/Resources/config.json
@@ -2,11 +2,14 @@
   "subjNum": 1,
   "subjSex": 0,
   "dataFile": "~\\..\\Assets\\TrialData\\sample_input.json",
-  "showFeedback": true,
-  "feedbackColor": "black",
   "targetCamera": true,
   "trackHeadPos": true,
   "initCameraX": 0,
   "initCameraY": 0,
-  "initCameraZ": 0
+  "initCameraZ": 0,
+  "canvasX": 0,
+  "canvasY": 6,
+  "canvasZ": 100,
+  "showFeedback": true,
+  "feedbackColor": "red"
 }
