+++ Assets/Scripts/ManageUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c122889481d56da45877b56063bbefe6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
